fix: 修复7项稳定性bug + 渡劫CD改为手动触发 + 数值问题文档
Critical bugs:
- GmService: isGm() 权限检查逻辑取反（GM被拒/普通用户可执行GM命令）
- EnhancementCore: matchesConstraints 重命名为 violatesConstraints（语义澄清）
- TravelService: travelTime 为 null 时 NPE
- Beast: calculateExpToNextLevel() level=0 返回0 导致 addExp 死循环
- CultivationService: LLM HTTP 调用在 @Transactional 内，改为先 save 后调用
- TrainingCombatLogic: skillsCache 永不刷新，改为直接查询
- UserStateService: loadUser 缺少 @Transactional，SELECT FOR UPDATE 无效

Functional:
- SpiritStoneService: withdraw/deposit 参数 int → long（对齐 User.spiritStones Long）
- TrainingService.endTraining: loadUser → loadUserForUpdate（消除 TOCTOU 竞态）
- Tribulation: 删除7天CD，isTribulationDue/calculateNextTribulationTime，改为福地手动触发
- Listener/Handler: 新增「福地渡劫」指令（OneBot+QQ双平台）

Docs:
- 新增 稳定性迭代计划.md，记录9项数值平衡问题

+++ src/main/java/top/stillmisty/xiantao/service/SpiritStoneService.java
@@ -16,7 +16,7 @@ public class SpiritStoneService {
   private final UserStateService userStateService;
 
   @Transactional
-  public void withdraw(Long userId, int amount) {
+  public void withdraw(Long userId, long amount) {
     if (amount <= 0) {
       throw new BusinessException(ErrorCode.PARAM_INVALID, "消耗灵石必须大于0");
     }
@@ -29,7 +29,7 @@ public class SpiritStoneService {
   }
 
   @Transactional
-  public void deposit(Long userId, int amount) {
+  public void deposit(Long userId, long amount) {
     if (amount <= 0) {
       throw new BusinessException(ErrorCode.PARAM_INVALID, "添加灵石必须大于0");
     }
